requirements.txt
torch>=1.7.1
torchvision>=0.8.2
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.56.0
clip11>=2.0.0
ftfy>=6.0.3
regex>=2021.4.4
scipy>=1.7.0
transformers>=4.11.0
matplotlib>=3.3.4